requirements.txt
Flask==3.0.0
reportlab==4.0.7
python-dotenv==1.0.0